refactor(mcp): make JetWhaleMcpCommand the only MCP plugin contract
The SDK is still alpha, so the legacy mcpTools()/handleMcpTool() pair is
removed instead of kept for compatibility. JetWhaleMcpCapablePlugin now
exposes just mcpCommands: List<JetWhaleMcpCommand>; McpToolRegistry derives
descriptors from the commands and renders JetWhaleMcpArgumentException as an
{"error": ...} payload at dispatch, which also removes the SDK's hand-rolled
JSON escaping. The example plugin and the server-service test fake migrate to
commands, and the network plugin's command classes move to one file each.
