+++ jetwhale-plugins/network/host/src/main/kotlin/com/kitakkun/jetwhale/plugins/network/host/GetTransactionCommand.kt
@@ -1,0 +1,29 @@
+package com.kitakkun.jetwhale.plugins.network.host
+
+import com.kitakkun.jetwhale.host.sdk.ExperimentalJetWhaleApi
+import com.kitakkun.jetwhale.host.sdk.JetWhaleMcpArgumentException
+import com.kitakkun.jetwhale.host.sdk.JetWhaleMcpArguments
+import com.kitakkun.jetwhale.host.sdk.JetWhaleMcpCommand
+import com.kitakkun.jetwhale.host.sdk.JetWhaleMcpParameterDescriptor
+
+@OptIn(ExperimentalJetWhaleApi::class)
+internal class GetTransactionCommand(
+    private val transactions: () -> List<HttpTransaction>,
+    private val redactForMcp: (HttpTransaction) -> HttpTransaction,
+) : JetWhaleMcpCommand() {
+    override val name = "$TOOL_PREFIX.getTransaction"
+    override val description = "Returns the full detail of one captured HTTP transaction (request/response headers and bodies, or the failure)."
+    override val parameters = mapOf(
+        "txId" to JetWhaleMcpParameterDescriptor(
+            type = "string",
+            description = "The transaction id from listTransactions.",
+        ),
+    )
+
+    override suspend fun execute(arguments: JetWhaleMcpArguments): String {
+        val txId = arguments.requireString("txId")
+        val transaction = transactions().firstOrNull { it.txId == txId }
+            ?: throw JetWhaleMcpArgumentException("no transaction with txId: $txId")
+        return redactForMcp(transaction).toDetailJson().toString()
+    }
+}
